change locator of logout button, move testcase to tests file
