+++ UI/Pages/MainPage.cs
@@ -13,7 +13,7 @@ namespace UI.Pages
         private static readonly By randomPageButton = By.CssSelector("#n-randompage>a>span");
         private static readonly By title = By.CssSelector("#Welcome_to_Wikipedia");
         private static readonly By userDropdownMenu = By.CssSelector("#vector-user-links-dropdown");
-        private static readonly By logoutButton = By.XPath("//*[@id='pt-logout']/a/span[2]");
+        private static readonly By logoutButton = By.Id("pt-logout");
         private static readonly By loggedUsername = By.CssSelector("#pt-userpage-2>a>span");
         private static readonly By aboutWikipediaLink = By.Id("n-aboutsite");
